Ch implememt jwt

+++ api/v1/views/redflags_views.py
@@ -1,15 +1,12 @@
 from flask import jsonify, make_response, request
+from api.v1.utilities.helpers import verify_admin, secured, verify_user
 from api.v1 import app
-from flask_jwt_extended import (
-    JWTManager,
-    jwt_required,
-    create_access_token,
-    get_jwt_identity)
 from api.v1.models import Redflags, Interventions
 from api.v1.utilities.helpers import update_redflags
 
 red = Redflags()
 incid = Interventions()
+
 
 @app.route('/')
 def home():
@@ -19,8 +16,12 @@ def home():
 
 
 @app.route('/api/v1/redflags', methods=['POST'])
-def create_redflag():
+@secured
+def create_redflag(logged_user):
     """ method implementing the create redflag api """
+    if verify_admin(logged_user):
+        return jsonify(
+            {'status': 403, 'error': "You do not have privileges to perform this request"}), 403
     try:
         data = request.get_json(force=True)
         redflag = {
@@ -39,7 +40,7 @@ def create_redflag():
                 {'status': 400, 'error': 'incident type can only be redflag'}), 400)
         elif not incid.valid_comment(redflag['comment']):
             return jsonify(
-            {'error': 'comment field should be a string and should contain alphabets'}), 400
+                {'error': 'comment field should be a string and should contain alphabets'}), 400
         elif not incid.validate_location(redflag['location']):
             return make_response(
                 jsonify(
@@ -61,7 +62,8 @@ def create_redflag():
 
 
 @app.route('/api/v1/redflags', methods=['GET'])
-def get_redflags():
+@secured
+def get_redflags(logged_user):
     """ method implementing get all redflags endpoint """
     all_redflags = red.get_redflags()
     if all_redflags:
@@ -70,7 +72,8 @@ def get_redflags():
 
 
 @app.route('/api/v1/redflags/<int:redflag_id>', methods=['GET'])
-def get_specific_redflag(redflag_id):
+@secured
+def get_specific_redflag(logged_user, redflag_id):
     """ gets a specific redflag by id """
     get_redflag = red.get_specific_redflag(redflag_id)
     if get_redflag:
@@ -82,8 +85,12 @@ def get_specific_redflag(redflag_id):
 
 
 @app.route('/api/v1/redflags/<int:redflag_id>/location', methods=['PATCH'])
-def edit_redflags_location(redflag_id):
+@secured
+def edit_redflags_location(logged_user, redflag_id):
     """ edits the location of a redflag """
+    if verify_admin(logged_user):
+        return jsonify(
+            {'status': 403, 'error': "You do not have privileges to perform this request"}), 403
     try:
         data = request.get_json(force=True)
 
@@ -97,8 +104,12 @@ def edit_redflags_location(redflag_id):
 
 
 @app.route('/api/v1/redflags/<int:redflag_id>/comment', methods=['PATCH'])
-def edit_redflags_comments(redflag_id):
+@secured
+def edit_redflags_comments(logged_user, redflag_id):
     """ edits the comments of a redflag """
+    if verify_admin(logged_user):
+        return jsonify(
+            {'status': 403, 'error': "You do not have privileges to perform this request"}), 403
     try:
         data = request.get_json(force=True)
 
@@ -106,33 +117,36 @@ def edit_redflags_comments(redflag_id):
             redflag_id, 'comment', data['comment'])
 
         return make_response(jsonify({"message": message}), status)
-    except:
+    except BaseException:
         return jsonify(
             {"error": "Please enter a valid key for comment field as comment"}), 400
 
 
 @app.route('/api/v1/redflags/<int:redflag_id>/status', methods=['PATCH'])
-@jwt_required
-def edit_redflags_status(redflag_id):
+@secured
+def edit_redflags_status(logged_user, redflag_id):
     """ edits the status of a redflag """
-    current_user = get_jwt_identity()
-    if current_user['role'] != 'admin':
+    if verify_user(logged_user):
         return jsonify(
-            {'message': 'Unauthorised access, please login as admin'}), 401
+            {'status': 403, 'error': "You do not have privileges to perform this request"}), 403
     try:
         data = request.get_json(force=True)
 
         message, status = update_redflags(redflag_id, 'status', data['status'])
 
         return make_response(jsonify({"message": message}), status)
-    except:
+    except BaseException:
         return jsonify(
             {"error": "Please enter a valid key for status field as status"}), 400
 
 
 @app.route('/api/v1/redflags/<int:redflag_id>', methods=['DELETE'])
-def del_redflag(redflag_id):
+@secured
+def del_redflag(logged_user, redflag_id):
     """ deletes a redflag from records """
+    if verify_admin(logged_user):
+        return jsonify(
+            {'status': 403, 'error': "You do not have privileges to perform this request"}), 403
     flag_list = red.get_specific_redflag(redflag_id)
     if flag_list:
         red.delete_redflag(redflag_id)
